+++ User/device/mpu6500.c
@@ -243,7 +243,7 @@ int8_t IMU_Parse(IMU_t *imu) {
 
   memcpy(raw, &(imu->raw), sizeof(imu->raw));
 
-  for (uint8_t i = 0; i < 20; i += 2)
+  for (size_t i = 0; i < 20; i += 2)
     ((int16_t *)&imu->raw)[i / 2] = (raw[i] << 8) | raw[i + 1];
 
   imu->data.temp = 21.0f + (float)imu->raw.temp / 333.87f;
